Delay acquiring the bus name until the engine is initialized
On Android, we initialize the control channel immediately and defer
initializing the engine until later. This is because on Android
the UI thread is blocked until the control channel is ready, because
of Java and Android's bindService threading models.

In GNOME, instead, the UI process uses asynchronous DBus to start
the service process, and never blocks. Which means that being
slow to acquire the DBus name is ok, as long as we don't timeout
(normal timeout is 15 seconds).
At the same time, if we initialize the name later, we are sure
that when we process requests from the UI process the engine will
be ready. This way, we avoid races at startup where randomly the UI will ask
to initialize the AssistantDispatcher before the engine is ready,
prompting the user to configure Matrix again.
This also matches the systemd semantics for DBus activated services.