Fix opening encrypted database created with a previous version
The previous version of sqlcipher used a different set of encryption
parameters, which would result in an unreadable database even with
the correct key.
Instead, tell sqlcipher which version to use, based on how
we created the database. If the database is initialized anew,
we will leave the default parameters, which are more secure.