Switch back to snowboy for wake-word detection
Mycroft Precise is unsustainable, it's just too big, too slow and
too messy to run. snowboy is easy to install, is lean, and supports
all the platforms we care about.

The wake word changes to "computer". It's better because it's
a universal model, that has seen a lot of speaker, compared to
the previous "almond" model which was mostly us (millenial male
non-native speakers). And "almond" is a bad hotword anyway.