shell-extension: remove the ability to hide the notification
Before 4979c898ec0da2cbe3e15c17ad83ac9d3b66fab2, we used
to override Notification.destroy to ignore it in some cases,
but that would reach a weird state where the notification
was not destroyed, but the NotificationMessage (the actual
widget in the calendar menu) was destroyed.

Instead, all we want is to hide the X button so the user doesn't
click it, and the rest follows.